requirements.txt
langchain 
chromadb
langchain-chroma
streamlit
google-generativeai 
langchain-google-genai 
python-dotenv 
pypdf